us #1123230 : allow sync branches to octane. fix duplications.
