+++ src/main/java/com/microfocus/application/automation/tools/octane/branches/BranchesPublisher.java
@@ -22,10 +22,8 @@ package com.microfocus.application.automation.tools.octane.branches;
 
 import com.cloudbees.plugins.credentials.CredentialsMatcher;
 import com.cloudbees.plugins.credentials.CredentialsMatchers;
-import com.cloudbees.plugins.credentials.CredentialsProvider;
 import com.cloudbees.plugins.credentials.common.StandardCredentials;
 import com.cloudbees.plugins.credentials.common.StandardUsernamePasswordCredentials;
-import com.cloudbees.plugins.credentials.domains.URIRequirementBuilder;
 import com.hp.octane.integrations.OctaneClient;
 import com.hp.octane.integrations.OctaneSDK;
 import com.hp.octane.integrations.exceptions.OctaneValidationException;
@@ -36,10 +34,7 @@ import com.hp.octane.integrations.services.pullrequestsandbranches.factory.Fetch
 import com.hp.octane.integrations.services.pullrequestsandbranches.factory.FetchHandler;
 import com.hp.octane.integrations.services.pullrequestsandbranches.rest.ScmTool;
 import com.hp.octane.integrations.services.pullrequestsandbranches.rest.authentication.AuthenticationStrategy;
-import com.hp.octane.integrations.services.pullrequestsandbranches.rest.authentication.BasicAuthenticationStrategy;
-import com.hp.octane.integrations.services.pullrequestsandbranches.rest.authentication.NoCredentialsStrategy;
-import com.hp.octane.integrations.services.pullrequestsandbranches.rest.authentication.PATStrategy;
-import com.microfocus.application.automation.tools.octane.GeneralUtils;
+import com.microfocus.application.automation.tools.octane.GitFetchUtils;
 import com.microfocus.application.automation.tools.octane.JellyUtils;
 import hudson.EnvVars;
 import hudson.Extension;
@@ -51,9 +46,7 @@ import hudson.tasks.BuildStepMonitor;
 import hudson.tasks.Publisher;
 import hudson.tasks.Recorder;
 import hudson.util.ListBoxModel;
-import hudson.util.Secret;
 import jenkins.tasks.SimpleBuildStep;
-import org.apache.commons.lang.StringUtils;
 import org.jenkinsci.Symbol;
 import org.jenkinsci.plugins.plaincredentials.StringCredentials;
 import org.kohsuke.stapler.AncestorInPath;
@@ -128,8 +121,8 @@ public class BranchesPublisher extends Recorder implements SimpleBuildStep {
 
         BranchFetchParameters fp = createFetchParameters(run, taskListener, logConsumer::printLog);
 
-        StandardCredentials credentials = getCredentialsById(myCredentialsId, run, taskListener.getLogger());
-        AuthenticationStrategy authenticationStrategy = getAuthenticationStrategy(credentials);
+        StandardCredentials credentials = GitFetchUtils.getCredentialsById(myCredentialsId, run, taskListener.getLogger());
+        AuthenticationStrategy authenticationStrategy = GitFetchUtils.getAuthenticationStrategy(credentials);
 
         try {
             //GET BRANCHES FROM CI SERVER
@@ -139,18 +132,18 @@ public class BranchesPublisher extends Recorder implements SimpleBuildStep {
             logConsumer.printLog("ALM Octane " + octaneClient.getConfigurationService().getConfiguration().geLocationForLog());
             octaneClient.validateOctaneIsActiveAndSupportVersion(PullRequestAndBranchService.BRANCH_COLLECTION_SUPPORTED_VERSION);
             BranchSyncResult result = OctaneSDK.getClientByInstanceId(myConfigurationId).getPullRequestAndBranchService()
-                    .syncBranchesToOctane(fetchHandler, fp, Long.parseLong(myWorkspaceId), GeneralUtils::getUserIdForCommit, logConsumer::printLog);
+                    .syncBranchesToOctane(fetchHandler, fp, Long.parseLong(myWorkspaceId), GitFetchUtils::getUserIdForCommit, logConsumer::printLog);
 
             BranchesBuildAction buildAction = new BranchesBuildAction(run, result, fp.getRepoUrl(), fp.getFilter());
             run.addAction(buildAction);
 
 
+        } catch (OctaneValidationException e) {
+            logConsumer.printLog("ALM Octane branch collector failed on validation : " + e.getMessage());
+            run.setResult(Result.FAILURE);
         } catch (Exception e) {
             logConsumer.printLog("ALM Octane branch collector failed : " + e.getMessage());
-            if(!(e instanceof OctaneValidationException)){
-                e.printStackTrace(taskListener.getLogger());
-            }
-
+            e.printStackTrace(taskListener.getLogger());
             run.setResult(Result.FAILURE);
         }
     }
@@ -219,23 +212,6 @@ public class BranchesPublisher extends Recorder implements SimpleBuildStep {
         }
     }
 
-    private AuthenticationStrategy getAuthenticationStrategy(StandardCredentials credentials) {
-        AuthenticationStrategy authenticationStrategy;
-        if (credentials == null) {
-            authenticationStrategy = new NoCredentialsStrategy();
-        } else if (credentials instanceof StringCredentials) {
-            Secret secret = ((StringCredentials) credentials).getSecret();
-            authenticationStrategy = new PATStrategy(secret.getPlainText());
-        } else if (credentials instanceof StandardUsernamePasswordCredentials) {
-            StandardUsernamePasswordCredentials cr = (StandardUsernamePasswordCredentials) credentials;
-            authenticationStrategy = new BasicAuthenticationStrategy(cr.getUsername(), cr.getPassword().getPlainText());
-        } else {
-            throw new IllegalArgumentException("Credentials type is not supported : " + credentials.getClass().getCanonicalName());
-        }
-
-        return authenticationStrategy;
-    }
-
     @Override
     public DescriptorImpl getDescriptor() {
         return (DescriptorImpl) super.getDescriptor();
@@ -260,25 +236,6 @@ public class BranchesPublisher extends Recorder implements SimpleBuildStep {
 
     public String getScmTool() {
         return scmTool;
-    }
-
-    /**
-     * Get user name password credentials by id.
-     */
-    private StandardCredentials getCredentialsById(String credentialsId, Run<?, ?> run, PrintStream logger) {
-
-        StandardCredentials credentials = null;
-        if (!StringUtils.isEmpty(credentialsId)) {
-            credentials = CredentialsProvider.findCredentialById(credentialsId,
-                    StandardCredentials.class,
-                    run,
-                    URIRequirementBuilder.create().build());
-            if (credentials == null) {
-                logger.println("Can not find credentials with the credentialsId:" + credentialsId);
-            }
-        }
-
-        return credentials;
     }
 
     @Symbol("collectBranchesToAlmOctane")
